[server] Now only administrator users can do administration actions

+++ server/src/diet/deleteMenu.js
@@ -11,20 +11,25 @@ export default (app) => {
     let menus;
     let menu;
 
+    if (req.user.role === false) {
+      res.status(403).send({error: 'Not enough rights to do this action!'});
+      return;
+    }
+
+    // check if user is the owner
+    if (req.user.id !== menu.owner) {
+      res.status(403).send({error: 'Not enough rights to do this action!!'});
+      return;
+    }
+
     try {
       menus = await Menu;
       menu = menus.filter( menu => menu.name === req.params.name ).reduce((a,b) => a.concat(b));
     } catch (e) {
-      res.status(400).send({error: 'Menu does not exist'});
-    }
-    
-    // check if user is the owner
-    if (req.user.id !== menu.owner) {
-      res.status(403).send({error: 'you can not entry !'});
-      return;
+      res.status(400).send({error: 'Menu not found'});
     }
 
-    // delete
+      // delete
     await menu.delete();
 
     // send success status
